This commit was generated by cvs2svn to compensate for changes in r5811, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5812 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/raster/raster-cache.h
@@ -140,7 +140,7 @@ raster_cache_data_fill (BYTE *dest,
       *xs = 0;
       *xe = length - 1;
       if (src_step == 1)
-	memcpy (dest, src, length);
+	memcpy(dest, src, (size_t)length);
       else
 	for (i = 0; i < length; i++, src += src_step)
 	  dest[i] = src[0];
@@ -366,7 +366,7 @@ raster_cache_data_fill_const(BYTE *dest,
       *xs = 0;
       *xe = length - 1;
       if (src_step == 1)
-        memset (dest, data, length);
+        memset(dest, data, (size_t)length);
       else
         for (i = 0; i < length; i++)
           dest[i] = data;
